docker-flask-redis-postgres-vol-compose/app/requirements.txt
Flask==2.3.2
redis==5.2.0
psycopg2-binary==2.9.7
